Wrapper: re-organize radspec postprocess to its own file
