+++ packages/aragon-wrapper/src/radspec/index.js
@@ -85,3 +85,5 @@ export async function tryDescribingUpgradeOrganizationBasket (intents, wrapper) 
     }
   }
 }
+
+export { postprocessRadspecDescription } from './postprocess'
